Standardize prime-number in go
